[FIX] Finish the migration to v11
