+++ account_mass_reconcile_by_purchase_line/README.rst
@@ -1,12 +1,12 @@
-.. image:: https://img.shields.io/badge/license-AGPLv3-blue.svg
-   :target: https://www.gnu.org/licenses/agpl.html
+.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/agpl
    :alt: License: AGPL-3
 
 =======================================
 Account Mass Reconcile by Purchase Line
 =======================================
 
-This module extends the functionality of acccount_mass_reconcile and
+This module extends the functionality of account_mass_reconcile and
 allow an user to reconcile debits and credits of an Account
 using the PO Line and Product as key fields. This type of
 reconciliation is to be used in the context of the Perpetual Inventory
@@ -20,7 +20,7 @@ To use this module, you need to:
 * Go to 'Accounting / Adviser / Mass Automatic Reconcile'.
 
 * Create a new reconciliation profile, and select a new configuration entry
-  with type 'Advanced. Product, purchase order line.'.
+  with type 'Advanced. Product, purchase order line'.
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
@@ -40,7 +40,7 @@ Credits
 Images
 ------
 
-* Odoo Community Association: `Icon <https://github.com/OCA/maintainer-tools/blob/master/template/module/static/description/icon.svg>`_.
+* Odoo Community Association: `Icon <https://odoo-community.org/logo.png>`_.
 
 Contributors
 ------------
