+++ account_mass_reconcile_by_purchase_line/models/__init__.py
@@ -1,3 +1,3 @@
 from . import mass_reconcile
 from . import base_advanced_reconciliation
-from . import simple_reconciliation
+from . import advanced_reconciliation
