+++ account_mass_reconcile_by_purchase_line/models/mass_reconcile.py
@@ -1,16 +1,17 @@
 # © 2015-18 Eficent Business and IT Consulting Services S.L. (www.eficent.com)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from odoo import api, models
+from odoo import models
 
 
 class AccountMassReconcileMethod(models.Model):
     _inherit = 'account.mass.reconcile.method'
 
-    @api.model
-    def _get_reconcilation_methods(self):
-        methods = super(AccountMassReconcileMethod,
-                        self)._get_reconcilation_methods()
+    @staticmethod
+    def _get_reconcilation_methods():
+        methods = super(
+            AccountMassReconcileMethod, AccountMassReconcileMethod).\
+            _get_reconcilation_methods()
         methods += [
             ('mass.reconcile.advanced.by.purchase.line',
              'Advanced. Product, purchase order line.'),
